+++ mitra-models/src/profiles/mod.rs
@@ -1,3 +1,4 @@
+mod checks;
 pub mod helpers;
 pub mod queries;
 pub mod types;